# Reduce allocations in UPath.Combine